Inventory backend code is done, needs testing and actual GUI implementation
I added Item as an abstract class as it works better for that type of
object and we can use polymorphism especially in the inventory. abstract
use() is in item therefore, consume and use as methods will not exist in
the item children. (well use() exists but anyways). I made a new scene
apparently when I started working. I added code to the assets. And shit
needs to be improved. The only thing that may not have been properly
implemented is the insert (2.4.2 because we have not yet gotten to GUIs
so I'll put that down in the sheet).